Implement proxy capabilities for postlicyd-rsyncrbl
The format for a configuration line is now

Destination=URL[proxy]

where the [proxy] part is not required if no proxy is needed. Carefully
analysing the script though, the ending ] is not strictly enforced, but
it helps understanding what's going on.

As whitespace now appears in the configuration, the parsing with "grep ="
is not valid anymore (as it would split the line). Therefore, the
configuration file now MUST have proper lines when they are not commented.
On the other hand, the new parsing makes it possible to have comments in
the end of the line as usual scripts.

A further change (in the script, does not affect the configuration file)
is that the order of the arguments to getfile has changed so that it
reflects more the configuration file instead of the rsync arguments.

	modified:   tools/postlicyd-rsyncrbl